scripts/requirements.txt
python-docx==1.1.2
pypdf>=6.1.3,<7
